requirements.txt
uvicorn
fastapi
fastapi-utilities
python-multipart
ruff
pydantic
pydantic[email]
pydantic-settings
pyjwt
passlib
psycopg2-binary
apscheduler
pytz
aiohttp
pytest
pytest-mock
pytest-asyncio
httpx